Updates to Brotli compression format, decoder and encoder
This commit contains a batch of changes that were made to the Brotli
compression algorithm in the last month. Most important changes:

   * Updated spec
   * Changed Huffman code length alphabet to use run length codes more
     efficiently, based on a suggestion by Robert Obryk
   * Changed encoding of the number of Huffman code lengths (HLEN)
   * Changed encoding of the number of Huffman trees (NTREES)
   * Added support for uncompressed meta-blocks

+++ enc/block_splitter.h
@@ -30,7 +30,7 @@ namespace brotli {
 struct BlockSplit {
   int num_types_;
   std::vector<uint8_t> types_;
-  std::vector<uint8_t> type_codes_;
+  std::vector<int> type_codes_;
   std::vector<int> lengths_;
 };
